+++ src/ArduinoJson/Strings/SizedFlashStringAdapter.hpp
@@ -15,15 +15,14 @@ class SizedFlashStringAdapter {
   SizedFlashStringAdapter(const __FlashStringHelper* str, size_t sz)
       : _str(str), _size(sz) {}
 
-  int8_t compare(const char* other) const {
+  int compare(const char* other) const {
     if (!other && !_str)
       return 0;
     if (!_str)
       return -1;
     if (!other)
       return 1;
-    return int8_t(
-        -strncmp_P(other, reinterpret_cast<const char*>(_str), _size));
+    return -strncmp_P(other, reinterpret_cast<const char*>(_str), _size);
   }
 
   bool equals(const char* expected) const {
